+++ src/app/app.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+
+import { Toolbar } from '@app/shared/components';
+
+@Component({
+  selector: 'app-root',
+  standalone: true,
+  imports: [RouterOutlet, Toolbar],
+  templateUrl: './app.html',
+  styleUrl: './app.scss',
+})
+export class App {}