+++ src/app/core/services/libraries-data.ts
@@ -5,9 +5,9 @@ import { Library } from '@app/core/interfaces';
 import { LIBRARIES } from '@data/libraries';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
-export class DataService {
+export class LibrariesData {
   private libraries$ = new BehaviorSubject<Library[]>(LIBRARIES);
 
   /**