+++ src/app/shared/components/library-card/library-card.ts
@@ -1,4 +1,4 @@
-import { Component, input, output } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input, output } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatRippleModule } from '@angular/material/core';
 
@@ -8,24 +8,25 @@ import { Library } from '@app/core/interfaces';
   selector: 'app-library-card',
   standalone: true,
   imports: [MatCardModule, MatRippleModule],
-  templateUrl: './library-card.component.html',
-  styleUrl: './library-card.component.scss'
+  templateUrl: './library-card.html',
+  styleUrl: './library-card.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class LibraryCardComponent {
+export class LibraryCard {
   readonly libraryCard = input<Library>({
     logoPath: '',
     name: '',
     description: '',
     link: '',
-    stepsToAdd: []
+    stepsToAdd: [],
   });
   readonly cardClick = output<Library>();
   readonly defaultImage: string = '/logos/image.svg';
 
   /**
-   * When card clicked, emits and event of current library card data
+   * Emits and event of current library card data
    */
-  clickCard(): void {
+  emitCardInfo(): void {
     this.cardClick.emit(this.libraryCard());
   }
 }