+++ CHANGELOG.md
@@ -9,15 +9,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 
-## [1.9.0] - 2025-12-12
+## [1.10.0] - 2026-03-24
 ### Added
+- On push change detection strategy for components
+### Changed
+- Angular update to v20
+- Migrate to zoneless change detection
+- Updates styles
+- Removed karma config and content
+- Removed modules
+- Files renamed according to new naming convention
+
+## [1.9.0] - 2025-12-12
 ### Changed
 - Angular update to v19
 - Migrate to inject()
 - Migrate inputs and outputs to signals
 - Toolbar style
 - Removed "public" access modifier from class fields
-### Fixed
 
 ## [1.8.2] - 2025-03-28
 ### Added